add look sensitivity setting
