+++ Assets/Scripts/UIAndMenus/PauseMenu/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2345e1f500a88342869b81cc672a62d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
