+++ Assets/Scripts/UIAndMenus/PauseMenu/Settings/SettingsSetups.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65aca24daeeda2b4da25e75f08b57229
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
